+++ CHANGELOG.md
@@ -7,17 +7,23 @@ and this project adheres to [Semantic Versioning].
 
 
 
+## [0.0.3] - 2024-05-21
+### Added
+- table sort
+- table search
+### Fixed
+- axios export
+- button props
+
+
 ## [0.0.2] - 2024-05-21
 
 ### Fixed
 - ltr regex
-
 ### Added
 - add table defaults
-
 ### Fixed
 - fix table issues 
-
 ## [0.0.1] - 2024-05-20
 
 - initial release